server/requirements.txt
# === Flask & Realtime Server ===
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
python-dotenv==1.0.1
eventlet==0.33.3
Werkzeug==2.3.7
simple-websocket==1.0.0
greenlet>=3.0.0
gunicorn==21.2.0  # ✅ ADDED - Required for Render

# === Database (Supabase) ===
supabase>=2.9.0

# === Core AI / OpenAI / LangChain ===
openai==1.109.1
langchain-core==1.0.1
langchain-openai==1.0.1
langsmith==0.4.38
tenacity==9.1.2
zstandard==0.25.0
packaging==24.2
pyyaml==6.0.2

# === Groq AI ===
groq==0.9.0  # ✅ ADDED - For Groq support

# === Data Handling / Utilities ===
regex==2024.9.11
requests==2.32.3
urllib3==2.2.2
certifi==2024.8.30

# === Audio & Speech (TTS/STT Support) ===
# pydub removed - incompatible with Python 3.13 (audioop module removed)
# STT endpoint will be disabled

# === Optional Tools (safe to keep) ===
anyio==4.11.0
httpx>=0.26,<0.29
distro==1.9.0
typing-extensions==4.12.2
colorama==0.4.6